+++ ryven/nodes/std/special_nodes.py
@@ -3,7 +3,7 @@ from contextlib import redirect_stdout, redirect_stderr
 
 from ryven.NENV import *
 widgets = import_widgets(__file__)
-from ryven.ironflow.node_interface.node_interface import SliderControl
+from ryven.ironflow.node_interface.control import SliderControl
 
 
 class NodeBase(Node):
